requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
segmentation-models-pytorch>=0.3.3
Pillow>=10.0.0
numpy>=1.24.0
